requirements.txt
# MultiHop Agent Requirements

# Core dependencies
pyyaml>=6.0

# Retrieval dependencies
rank_bm25>=0.2.2
torch>=1.12.0
transformers>=4.20.0

# Graph database
neo4j>=4.4.0

# Web automation
selenium>=4.0.0
beautifulsoup4>=4.10.0

# Utilities
requests>=2.28.0
tqdm>=4.64.0
loguru>=0.6.0

# For Contriever (if used)
faiss-cpu>=1.7.0
